requirements.txt
# 核心依赖
mcp>=1.6.0
openai>=1.76.0
openai-agents>=0.0.13
python-dotenv>=1.1.0
requests>=2.32.3
aiohttp>=3.11.18
pytz>=2024.1
colorama>=0.4.6
python-dateutil>=2.9.0.post0

# GRAG知识图谱相关依赖
py2neo>=2021.2.3
pyvis>=0.3.2
matplotlib>=3.10.0

# API服务器相关依赖
flask-cors>=6.0.1
flask>=3.1.0
gevent>=25.5.1
fastapi>=0.115.0

# 网络通信相关依赖
uvicorn[standard]>=0.34.0
librosa
websockets>=14.0

# AI机器学习相关依赖
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
tqdm>=4.67.1
scikit-learn>=1.6.1
scipy>=1.15.2
transformers>=4.51.3

# 数据处理相关依赖
pydantic>=2.11.3
pydantic-settings>=2.9.1
griffe>=1.7.3
anyio>=4.9.0
httpx>=0.28.1
httpx-sse>=0.4.0
sse-starlette>=2.3.3
starlette>=0.46.2
certifi>=2025.4.26
charset-normalizer>=3.4.1
idna>=3.10
urllib3>=2.4.0
typing-extensions>=4.13.2
markdown>=3.8

# GUI界面相关依赖
pyqt5-qt5==5.15.2
pyqt5==5.15.11
playwright>=1.52.0
greenlet>=3.2.1
pyee>=13.0.0
pygame>=2.6.0
html2text>=2020.1.16

# 音频处理相关依赖
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudio>=0.2.14
edge-tts>=7.0.2
emoji>=2.8.0

# 语音识别相关依赖
winrt-runtime>=3.2.1

# 系统托盘相关依赖
pystray>=0.19.4
pillow>=10.0.0

# 其他工具依赖
tiktoken
bottleneck>=1.3.6

# 系统控制相关依赖
screen-brightness-control
pycaw
comtypes

# MCP工具相关依赖
jmcomic>=2.6.5
fastmcp>=2.11.2

# MQTT通信相关依赖
paho-mqtt>=2.0.0

# 其他依赖
python-docx>=1.2.0
mqtt-tool>=0.1.1
langchain_community
langchain